service: tweak generated ipsec configuration
The ipsec %defaultroute keyword is used to have Strongswan use the
local address of the default route network interface at the point of
startup or configuration reload.  It's probably a fair bet that this
would be the interface used to connect to the VPN gateway, but that's
not guaranteed.  To support the latter, determine the IP address
directly by connecting to the gateway and using the getsockname(2)
system call to derive the local socket address.

The left|rightprotoport options have been deprecated since Strongswan
5.1.0 which was released in 2013.  Update to the currently supported
option left|rightsubnet.